viewer/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg[binary]>=3.1,<3.3
humanize==4.10.0
requests==2.32.3
urllib3==2.2.3
idna==3.10
charset-normalizer==3.4.3
certifi==2025.8.3
gunicorn==22.0.0
